+++ persistent-mysql/ChangeLog.md
@@ -1,3 +1,9 @@
+# Changelog for persistent-mysql
+
+## 2.9.2
+
+* Fix [832](https://github.com/yesodweb/persistent/issues/832): `repsertMany` now matches `mapM_ (uncurry repsert)` and is atomic.
+
 ## 2.9.1
 
 * Added support for SQL isolation levels to via SqlBackend. [#812]